naming of async methods have been changed, getCount changed from int to long, unnecessary namings have been removed.

+++ iTechArt.Domain/ModelInterfaces/IGraph.cs
@@ -11,17 +11,17 @@ namespace iTechArt.Domain.ModelInterfaces
         /// <summary>
         /// Gets or sets the males count in table.
         /// </summary>
-        public double MaleAmount { get; set; }
+        public int MaleAmount { get; set; }
 
         /// <summary>
         /// Gets or sets the females count in table.
         /// </summary>
-        public double FemaleAmount { get; set; }
+        public int FemaleAmount { get; set; }
 
         /// <summary>
         /// Gets or sets the average age count of males in table.
         /// </summary>
-        public int AverageAgeMale { get; set; }
+        public double AverageAgeMale { get; set; }
 
         /// <summary>
         /// Gets or sets the average age count of females in table.
